#69: Moved New Assessment Button to Assessments Page, hiding it if Assessment already in progress
