+++ React-Frontend/src/containers/assessments.js
@@ -1,14 +1,47 @@
 import React, { Component } from 'react'
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux'
-import { fetchAssessments, selectAssessment } from '../actions/index';
+import { fetchAssessments, selectAssessment, postAssessment, ROOT_URL } from '../actions/index';
 import { Link } from 'react-router';
-
+import axios from 'axios';
 
 class Assessments extends Component {
 
   componentWillMount() {
 		this.props.fetchAssessments(this.props.login.user_id);
+  }
+
+  addAssessment() {
+    axios.get(`${ROOT_URL}versions`)
+    .then((response) => {
+      let latest = 0.0;
+      let latest_id = null;
+      response.data.rows.map((data) => {
+        let current = parseFloat(data.version);
+        if (current > latest) {
+          latest = current;
+          latest_id = data.version_id;
+        }
+      })
+      this.props.postAssessment(this.props.login.user_id, latest_id);
+	  })
+	  .catch((error) => {
+		  console.log(error);
+	  });
+  }
+
+  renderNewAssessment(assessments) {
+      let inProgress = false;
+      for (var i in assessments) {
+        if (assessments[i].end_time == null) {
+            inProgress = true;
+        }
+      }
+      if (!inProgress) { // Hide the Add new Assessment Link if there is one in progress
+          return (
+            <Link to={"/category/1"} className="home-screen-button" onClick={() => this.addAssessment()}>Begin a New Project Assessment</Link>
+          )
+      }
   }
 
   renderAssessment(data) {
@@ -44,6 +77,7 @@ class Assessments extends Component {
                         <div className="col-md-3 home-screen-button-container">
                             <p>Assessments for { this.props.login.team_name }</p>
                             { assessments.map(this.renderAssessment, this) }
+                            { this.renderNewAssessment(assessments) }
                         </div>
                     </div>
                 </div>
@@ -56,4 +90,4 @@ function mapStateToProps(state) {
 	return { assessments: state.assessments, login: state.login };
 }
 
-export default connect(mapStateToProps, { fetchAssessments,selectAssessment })(Assessments);
+export default connect(mapStateToProps, { fetchAssessments,selectAssessment,postAssessment })(Assessments);
